Pipe demo
proposal for grid based level building. this demo shows basic player movement and the mechanics pipe building ability.

this model was developed in order to make creating and testing levels easier for those without a focus in programming. a gameobject brush can be used to place objects into the scene in a grid pattern from existing prefabs.